+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added parsing command line arguments.
 - Added loading parameters from a Python file.
 - Added saving parameters to a JSON file.
+- Added saving software versions to a JSON file.
 - Added hierarchy of SimTools exceptions.
 - Added package installer.
 - Added package description.